refactored save_refill into medication_refill_helper
