+++ app/controllers/medications_controller.rb
@@ -5,7 +5,7 @@ require 'google/api_client'
 class MedicationsController < ApplicationController
   include CollectionPageSetup
   include ReminderHelper
-  helper_method :save_refill_to_google_calendar
+  include MedicationRefillHelper
   before_action :set_medication, only: %i[show edit update destroy]
 
   # GET /medications
@@ -81,31 +81,13 @@ class MedicationsController < ApplicationController
     end
   end
 
-  # Save refill date to Google calendar
-  def save_refill_to_google_calendar(medication)
-    return true unless current_user.google_oauth2_enabled? &&
-                       medication.add_to_google_cal == '1'
-
-    summary = 'Refill for ' + medication.name
-    begin
-      CalendarUploader.new(summary: summary,
-                           date: medication.refill,
-                           access_token: current_user.access_token,
-                           email: current_user.email).upload_event
-    rescue
-      return_to_sign_in
-      false
-    else
-      true
-    end
-  end
-
   def return_to_sign_in
     sign_out current_user
     respond_to do |format|
       format.html { redirect_to new_user_session_path }
       format.json { head :no_content }
     end
+    false
   end
 
   private
